+++ plugin-packs/postcss-preset-env/src/plugins/plugins-by-id.mjs
@@ -6,6 +6,7 @@ import postcssCascadeLayers from '@csstools/postcss-cascade-layers';
 import postcssAttributeCaseInsensitive from 'postcss-attribute-case-insensitive';
 import postcssClamp from 'postcss-clamp';
 import postcssColorFunction from '@csstools/postcss-color-function';
+import postcssColorFunctionDisplayP3Linear from '@csstools/postcss-color-function-display-p3-linear';
 import postcssColorFunctionalNotation from 'postcss-color-functional-notation';
 import postcssColorMixFunction from '@csstools/postcss-color-mix-function';
 import postcssColorMixVariadicFunctionArguments from '@csstools/postcss-color-mix-variadic-function-arguments';
@@ -72,6 +73,7 @@ export const pluginsById = new Map(
 		['case-insensitive-attributes', postcssAttributeCaseInsensitive],
 		['clamp', postcssClamp],
 		['color-function', postcssColorFunction],
+		['color-function-display-p3-linear', postcssColorFunctionDisplayP3Linear],
 		['color-functional-notation', postcssColorFunctionalNotation],
 		['color-mix', postcssColorMixFunction],
 		['color-mix-variadic-function-arguments', postcssColorMixVariadicFunctionArguments],